+++ src/main/java/com/solvd/blog/mapper/impl/UserMapper.java
@@ -2,7 +2,7 @@ package com.solvd.blog.mapper.impl;
 
 import com.solvd.blog.mapper.Mapper;
 import com.solvd.blog.model.User;
-import com.solvd.blog.neo4j.NeoUser;
+import com.solvd.blog.neo4j.NjUser;
 import org.neo4j.driver.Record;
 import org.neo4j.driver.types.Node;
 import org.springframework.stereotype.Component;
@@ -15,7 +15,7 @@ public class UserMapper implements Mapper<User> {
     @Override
     public User toEntity(final Record record) {
         Node node = record.get("user").asNode();
-        return new NeoUser(
+        return new NjUser(
                 node.id(),
                 node.get("name").asString(),
                 node.get("email").asString(),
