+++ src/main/java/com/solvd/blog/serializer/PostSerializer.java
@@ -3,15 +3,15 @@ package com.solvd.blog.serializer;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
-import com.solvd.blog.neo4j.NeoPost;
+import com.solvd.blog.neo4j.NjPost;
 
 import java.io.IOException;
 
-public class PostSerializer extends JsonSerializer<NeoPost> {
+public class PostSerializer extends JsonSerializer<NjPost> {
 
     @Override
     public void serialize(
-            final NeoPost value,
+            final NjPost value,
             final JsonGenerator gen,
             final SerializerProvider serializers
     ) throws IOException {
