+++ src/main/java/com/solvd/blog/transaction/TxUsers.java
@@ -1,15 +1,15 @@
-package com.solvd.blog.unique;
+package com.solvd.blog.transaction;
 
-import com.solvd.blog.exception.ResourceAlreadyExistsException;
 import com.solvd.blog.model.User;
 import com.solvd.blog.model.Users;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Component
-public class UqUsers implements Users {
+public class TxUsers implements Users {
 
     private final Users users;
 
@@ -18,36 +18,36 @@ public class UqUsers implements Users {
      *
      * @param users Users
      */
-    public UqUsers(@Qualifier("txUsers") final Users users) {
+    public TxUsers(@Qualifier("njUsers") final Users users) {
         this.users = users;
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<User> iterate() {
         return this.users.iterate();
     }
 
     @Override
+    @Transactional(readOnly = true)
     public User user(final Long id) {
         return this.users.user(id);
     }
 
     @Override
+    @Transactional
     public User add(final User user) {
-        if (this.exists(user.email())) {
-            throw new ResourceAlreadyExistsException(
-                    "Email should be unique!"
-            );
-        }
         return this.users.add(user);
     }
 
     @Override
+    @Transactional
     public User update(final User user) {
         return this.users.update(user);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Boolean exists(final String email) {
         return this.users.exists(email);
     }
