+++ src/main/java/com/solvd/blog/validation/VdUsers.java
@@ -1,15 +1,15 @@
-package com.solvd.blog.tx;
+package com.solvd.blog.validation;
 
+import com.solvd.blog.exception.ResourceAlreadyExistsException;
 import com.solvd.blog.model.User;
 import com.solvd.blog.model.Users;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Component
-public class TxUsers implements Users {
+public class VdUsers implements Users {
 
     private final Users users;
 
@@ -18,36 +18,36 @@ public class TxUsers implements Users {
      *
      * @param users Users
      */
-    public TxUsers(@Qualifier("neoUsers") final Users users) {
+    public VdUsers(@Qualifier("txUsers") final Users users) {
         this.users = users;
     }
 
     @Override
-    @Transactional(readOnly = true)
     public List<User> iterate() {
         return this.users.iterate();
     }
 
     @Override
-    @Transactional(readOnly = true)
     public User user(final Long id) {
         return this.users.user(id);
     }
 
     @Override
-    @Transactional
     public User add(final User user) {
+        if (this.exists(user.email())) {
+            throw new ResourceAlreadyExistsException(
+                    "Email should be unique!"
+            );
+        }
         return this.users.add(user);
     }
 
     @Override
-    @Transactional
     public User update(final User user) {
         return this.users.update(user);
     }
 
     @Override
-    @Transactional(readOnly = true)
     public Boolean exists(final String email) {
         return this.users.exists(email);
     }
