+++ src/main/java/com/solvd/blog/web/UserController.java
@@ -36,7 +36,7 @@ public final class UserController {
      * @param posts TxPosts
      */
     public UserController(
-            @Qualifier("uqUsers") final Users users,
+            @Qualifier("vdUsers") final Users users,
             @Qualifier("txPosts") final Posts posts
     ) {
         this.users = users;
